ssh+blake2b+tests: tests passing everywhere, ssh generation perfect

+++ include/wasm.h
@@ -25,7 +25,5 @@
 
 #define STRLEN_CONST(v) (sizeof(v "") - 1)
 
-#include <stdint.h>
-
 int printf(const char *restrict format, ...) __attribute__((format(printf, 1, 2)));
-int snprintf(char *restrict str, uint32_t size, const char *restrict format, ...) __attribute__((format(printf, 3, 4)));
+int snprintf(char *restrict str, unsigned long size, const char *restrict format, ...) __attribute__((format(printf, 3, 4)));
